requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
Django==3.2.15
djangorestframework==3.13.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
